move form config for visa logic
