+++ src/app/[locale]/calculator/[visa]/[section]/[prompt]/page.tsx
@@ -1,5 +1,5 @@
 import { VisaForm } from '@components/VisaForm'
-import { formConfigForVisa } from '@lib/domain/visa'
+import { formConfigForVisa } from '@lib/domain/form'
 import { notFound } from 'next/navigation'
 
 interface Props {
